Add weather back into the airport class once again
