+++ lib/airport.rb
@@ -1,5 +1,4 @@
 require_relative 'plane'
-require_relative 'weather'
 class Airport
 
   AIRPORT_CAPACITY = 20
@@ -13,16 +12,16 @@ class Airport
 
   public
 
-  def land(plane, weather)
-    fail "Bad weather today. Cannot land." if weather.stormy_weather?
+  def land(plane)
+    fail "Bad weather today. Cannot land." if stormy_weather?
     fail "Sorry. Airport full. Go away." if full?
     fail "This plane is already landed." if plane.landed?
     plane.land_plane
     planes << plane
   end
 
-  def take_off(plane, weather)
-    fail "Bad weather today. Cannot take off." if weather.stormy_weather?
+  def take_off(plane)
+    fail "Bad weather today. Cannot take off." if stormy_weather?
     fail "Plane is already in the air!" if !plane.landed?
     plane.take_off_plane
     planes.delete(plane)
@@ -30,6 +29,15 @@ class Airport
 
   def set_capacity(number)
     @capacity = number
+  end
+
+  def stormy_weather?
+    random_number = rand(10)
+    if random_number >= 8
+      true
+    else
+      false
+    end
   end
 
   private
